Make sorting of paycheck postings configurable
If the `sort_postings` option in the importer config is set to `False`,
then related postings are left next to each other. The option defaults
to `True` to match the existing behavior.

For example, with a configuration containing a template such as:

```
'Employer Paid Benefits': {
    '401(k) Employer Match': [
	'Income:Hooli:Benefits:401k-Match',
	'Assets:Hooli-401k-Match',
    ],
    ...
 ```

It's helful if the generated beancount transaction includes the
`Income:Hooli:Benefits:401k-Match` and `Assets:Hooli-401k-Match`
postings on consecutive lines rather than intermixed with postings from
other paycheck line items.