Importer.extract: Explicitly name args to Transaction
This shows that payee and narration are switched, as documented.
https://github.com/beancount/beancount/blob/v2/beancount/core/data.py is
the reference. This confused my use of smart_importer completion from an
existing ledger and took me some time to figure out.

I think the arguments should be swithed but I can see this would be
difficult to do while maintaining backwards compatibility.